+++ Changelog.md
@@ -48,6 +48,7 @@
 * Add parsing `DocxParagraphRun#object` and `Object#ole_object`
 * New way to parse default RunProperties and ParagraphProperties. Old way is still there
 * Add parsing `RunStyle`. Implement #140
+* Add parsing `Chart#view3D`
 
 ### Fixes
 * Fix parsing document style id - it can be string, not only digit